# beryl-core-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beryl Core API",
    "description": "Central API Gateway and orchestration layer for Béryl ecosystem",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/social/feed/personalized": {
      "post": {
        "tags": [
          "Social"
        ],
        "summary": "Get Personalized Feed",
        "description": "Get personalized feed for user.",
        "operationId": "get_personalized_feed_api_v1_social_feed_personalized_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonalizedFeedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedItemResponse"
                  },
                  "type": "array",
                  "title": "Response Get Personalized Feed Api V1 Social Feed Personalized Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/social/recommendations": {
      "post": {
        "tags": [
          "Social"
        ],
        "summary": "Get Recommendations",
        "description": "Get AI recommendations for user.",
        "operationId": "get_recommendations_api_v1_social_recommendations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecommendationResponse"
                  },
                  "type": "array",
                  "title": "Response Get Recommendations Api V1 Social Recommendations Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/social/behavior/analyze": {
      "post": {
        "tags": [
          "Social"
        ],
        "summary": "Analyze Behavior",
        "description": "Analyze user behavior patterns.",
        "operationId": "analyze_behavior_api_v1_social_behavior_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BehaviorAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BehaviorAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/social/moderation/check": {
      "post": {
        "tags": [
          "Social"
        ],
        "summary": "Check Moderation",
        "description": "Check content for moderation flags.",
        "operationId": "check_moderation_api_v1_social_moderation_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerationCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/social/trending": {
      "get": {
        "tags": [
          "Social"
        ],
        "summary": "Get Trending",
        "description": "Get trending content.",
        "operationId": "get_trending_api_v1_social_trending_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrendingContentResponse"
                  },
                  "title": "Response Get Trending Api V1 Social Trending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/social/sentiment/community": {
      "post": {
        "tags": [
          "Social"
        ],
        "summary": "Analyze Sentiment",
        "description": "Analyze community sentiment.",
        "operationId": "analyze_sentiment_api_v1_social_sentiment_community_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SentimentAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SentimentAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/social/connections/suggest/{user_id}": {
      "get": {
        "tags": [
          "Social"
        ],
        "summary": "Suggest Connections",
        "description": "Suggest user connections.",
        "operationId": "suggest_connections_api_v1_social_connections_suggest__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionSuggestionResponse"
                  },
                  "title": "Response Suggest Connections Api V1 Social Connections Suggest  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BehaviorAnalysisRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "User identifier"
          },
          "period_days": {
            "type": "integer",
            "title": "Period Days",
            "description": "Analysis period",
            "default": 30
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "BehaviorAnalysisRequest",
        "description": "Request for behavior analysis."
      },
      "BehaviorAnalysisResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "engagement_level": {
            "type": "string",
            "title": "Engagement Level"
          },
          "preferred_content_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Content Types"
          },
          "posting_frequency": {
            "type": "number",
            "title": "Posting Frequency"
          },
          "interaction_patterns": {
            "type": "object",
            "title": "Interaction Patterns"
          },
          "peak_activity_hours": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Peak Activity Hours"
          },
          "community_affinity": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Community Affinity"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "engagement_level",
          "preferred_content_types",
          "posting_frequency",
          "interaction_patterns",
          "peak_activity_hours",
          "community_affinity",
          "timestamp"
        ],
        "title": "BehaviorAnalysisResponse",
        "description": "Response for behavior analysis."
      },
...
      "ConnectionSuggestionResponse": {
        "properties": {
          "suggestion_id": {
            "type": "string",
            "title": "Suggestion Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "suggested_user_id": {
            "type": "string",
            "title": "Suggested User Id"
          },
          "relevance_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Relevance Score"
          },
          "common_interests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Common Interests"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "suggestion_id",
          "user_id",
          "suggested_user_id",
          "relevance_score",
          "common_interests",
          "timestamp"
        ],
        "title": "ConnectionSuggestionResponse",
        "description": "Response for connection suggestion."
      },
...
      "FeedItemResponse": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "author_id": {
            "type": "string",
            "title": "Author Id"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "content_text": {
            "type": "string",
            "title": "Content Text"
          },
          "engagement_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Engagement Score"
          },
          "predicted_engagement": {
            "type": "number",
            "title": "Predicted Engagement"
          },
          "rank_score": {
            "type": "number",
            "title": "Rank Score"
          },
          "timestamp_created": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp Created"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "author_id",
          "content_type",
          "content_text",
          "engagement_score",
          "predicted_engagement",
          "rank_score",
          "timestamp_created"
        ],
        "title": "FeedItemResponse",
        "description": "Response for feed item."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModerationCheckRequest": {
        "properties": {
          "content_id": {
            "type": "string",
            "title": "Content Id",
            "description": "Content identifier"
          },
          "content_text": {
            "type": "string",
            "title": "Content Text",
            "description": "Content to check"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "description": "Type of content",
            "default": "post"
          }
        },
        "type": "object",
        "required": [
          "content_id",
          "content_text"
        ],
        "title": "ModerationCheckRequest",
        "description": "Request for content moderation."
      },
      "ModerationCheckResponse": {
        "properties": {
          "content_id": {
            "type": "string",
            "title": "Content Id"
          },
          "risk_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score"
          },
          "violation_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Violation Types"
          },
          "confidence": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "recommended_action": {
            "type": "string",
            "title": "Recommended Action",
            "description": "allow, review, remove, block"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "content_id",
          "risk_score",
          "violation_types",
          "confidence",
          "recommended_action",
          "timestamp"
        ],
        "title": "ModerationCheckResponse",
        "description": "Response for moderation check."
      },
...
      "PersonalizedFeedRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "User identifier"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Number of items",
            "default": 20
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Pagination offset",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "PersonalizedFeedRequest",
        "description": "Request for personalized feed."
      },
...
      "RecommendationRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "User identifier"
          },
          "recommendation_type": {
            "type": "string",
            "title": "Recommendation Type",
            "description": "content, users, communities",
            "default": "content"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Number of recommendations",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "RecommendationRequest",
        "description": "Request for recommendations."
      },
      "RecommendationResponse": {
        "properties": {
          "recommendation_id": {
            "type": "string",
            "title": "Recommendation Id"
          },
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "recommendation_type": {
            "type": "string",
            "title": "Recommendation Type"
          },
          "confidence_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confidence Score"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "recommendation_id",
          "target_id",
          "recommendation_type",
          "confidence_score",
          "reason",
          "timestamp"
        ],
        "title": "RecommendationResponse",
        "description": "Response for recommendation."
      },
...
      "SentimentAnalysisRequest": {
        "properties": {
          "community_id": {
            "type": "string",
            "title": "Community Id",
            "description": "Community identifier"
          },
          "period_days": {
            "type": "integer",
            "title": "Period Days",
            "description": "Analysis period",
            "default": 7
          }
        },
        "type": "object",
        "required": [
          "community_id"
        ],
        "title": "SentimentAnalysisRequest",
        "description": "Request for sentiment analysis."
      },
      "SentimentAnalysisResponse": {
        "properties": {
          "community_id": {
            "type": "string",
            "title": "Community Id"
          },
          "overall_sentiment": {
            "type": "string",
            "title": "Overall Sentiment",
            "description": "positive, neutral, negative"
          },
          "sentiment_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": -100.0,
            "title": "Sentiment Score"
          },
          "emotion_breakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Emotion Breakdown"
          },
          "key_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Topics"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "community_id",
          "overall_sentiment",
          "sentiment_score",
          "emotion_breakdown",
          "key_topics",
          "timestamp"
        ],
        "title": "SentimentAnalysisResponse",
        "description": "Response for sentiment analysis."
      },
...
      "TrendingContentResponse": {
        "properties": {
          "content_id": {
            "type": "string",
            "title": "Content Id"
          },
          "author_id": {
            "type": "string",
            "title": "Author Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "trend_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Trend Score"
          },
          "engagement_count": {
            "type": "integer",
            "title": "Engagement Count"
          },
          "growth_rate": {
            "type": "number",
            "title": "Growth Rate"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "content_id",
          "author_id",
          "title",
          "category",
          "trend_score",
          "engagement_count",
          "growth_rate",
          "timestamp"
        ],
        "title": "TrendingContentResponse",
        "description": "Response for trending content."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}